requirements.txt
pyrogram
tgcrypto
py-tgcalls==0.5.5
python-dotenv
yt-dlp
youtube_dl
youtube_search_python
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg
Pillow
future
psutil
wget
tswift

